+++ NEWS.md
@@ -2,5 +2,10 @@
 
 * Implemented speed permutation-based statistical tests (one sample and two samples)
 
+## version 0.1.1 [2020-10-11]
 
-
+* Add pecora-package.md
+* Changed data matrix X dimension
+* Changed documentation and rand arguments
+* Deleted 
+* Created separted function to compute p-values